refactor warnings
