--- packages/config-plugins/src/__tests__/Serialize-test.ts
@@ -1,19 +1,0 @@
-import { serializeAndEvaluate } from '../Serialize';
-
-describe('serializeAndEvaluate', () => {
-  it(`serializes item`, () => {
-    expect(
-      serializeAndEvaluate({
-        foo: 'bar',
-        boo: true,
-        inn: 200,
-        then: [true, { foo: 'bar' }],
-        alpha: () => ({ beta: ['val'] }),
-        last: {
-          bar: 'foo',
-          charlie: [2, 'delta'],
-        },
-      })
-    ).toMatchSnapshot();
-  });
-});
