--- packages/config-plugins/src/__tests__/fixtures/dynamic-cwd/app.config.ts
@@ -1,10 +1,0 @@
-import { ConfigContext, ExpoConfig } from '@expo/config';
-
-export default ({ config }: ConfigContext): ExpoConfig => {
-  const expoConfig = {
-    extra: {
-      processCwd: process.cwd(),
-    },
-  };
-  return expoConfig;
-};
