--- packages/config-plugins/src/__tests__/fixtures/require-file/export-json_app.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  foo: 'bar',
-  name: 'cool+export-json_app.config',
-};
