--- packages/config-plugins/src/__tests__/fixtures/require-file/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "js-config-test",
-  "version": "1.0.0"
-}
