--- packages/config-plugins/src/__tests__/fixtures/require-file/with-import_app.config.js
@@ -1,6 +1,0 @@
-const { foo } = require('./export-json_app.config');
-
-module.exports = function ({ config }) {
-  config.foo = foo;
-  return config;
-};
