+++ packages/config-plugins/src/utils/errors.ts
