+++ packages/config/src/index.ts
@@ -1,6 +1,5 @@
 /* eslint-disable import/export */
-import * as WarningAggregator from './WarningAggregator';
-export { WarningAggregator };
+
 export * from './Modules';
 export * from './Config';
 export * from './Config.types';
