+++ packages/expo-cli/src/commands/utils/logConfigWarnings.ts
@@ -1,4 +1,4 @@
-import { WarningAggregator } from '@expo/config';
+import { WarningAggregator } from '@expo/config-plugins';
 import chalk from 'chalk';
 
 import log from '../../log';
